Tile Coloring and Android Game Settings
Made changes to tile coloring and solution checking. Solution checking is still broken, but is better than it was. Still seems to be that  3-way and 4-way pipes are what fuck up the solution checking.

Also made changes to project settings and tested it on Android, since this will eventually be an Android game. I set it so that it can only be played in landscape mode.

TODO:
- Hook into Android back-button events to back out of game properly. Right now it does nothing.

- Pipe tiles seem too small, especially for the first levels. Maybe save the small tiles for later, harder levels and make simpler/bigger tilesets for the early levels.

- Need Android icon.

- Need game company names. So far the stand-in is MyKrappyGames.