DYNAMIC SOLUTION CHECKING WORKS
:D I did it!

Just had to remove the 'return false' statements, replace them with continues, and add a boolean indicating whether leaks have been found.  Also had to move the check for the EndPipe from the FindSolution function into the CheckIfSolutionFound function, since not all branches would contain the EndPipe and that was causing it to fail.

Also made it so that coloring only happens if a debug bool is set in PipeFaceCalculator, and that pipe coloring is reset each time before being reapplied, which makes things a lot clearer.

Need to rename FindSolution to something more appropriate, like CheckNoLeaksExist or something.

Need to rename PipeFaceCalculator to represent its new all-encompassing nature.

Need to adjust it so that "open" end pipes count as a leak if they are in the solution. Probably just need to add a "if (currentTile == "OpenDeadEnd") foundLeaks = true;" statement in the FindSolution() check to piggy back off the leaks boolean that I just put in.

Need to go back and clean out a lot of commented code too.